CMakeLists: Fix Windows symlink issues
There are permission issues on Windows when creating symlinks and
hard links. I've tested all permutations, and the only consistent
method was to use junctions. CMake has the flags for creating
junctions in its core, but they are not exposed anywhere. I will try
to submit a patch there to get that corrected, but for now this
change should allow CMake to function reliably on Windows.