JaktTest: Build directory validation
This change does a test for the existence of the jakt binary before
proceeding to the test run. Without this, if there was an error in the
build path or if the build was not at the default location, jakttest
would start anyway and begin dumping masses of python errors.